move form to components because of usage from more places
